Pulled settings persistance out from the log file reader file
